getIndices method -> getIndicesOf helper function
